+++ lib/rust/ensogl/component/grid-view/src/selectable/highlight.rs
@@ -4,9 +4,10 @@
 use crate::prelude::*;
 
 use crate::entry;
-use crate::entry_position;
-use crate::selectable::highlight::shape::HoverAttrSetter;
-use crate::selectable::highlight::shape::SelectionAttrSetter;
+use crate::header;
+use crate::selectable::highlight::connected_entry::ConnectedEntry;
+use crate::selectable::highlight::connected_entry::EndpointsGetter;
+use crate::selectable::highlight::layer::HasConstructor as TRAIT_HasConstructor;
 use crate::Col;
 use crate::Entry;
 use crate::Row;
@@ -21,8 +22,26 @@ use ensogl_core::Animation;
 // === Export ===
 // ==============
 
+pub mod connected_entry;
 pub mod layer;
 pub mod shape;
+
+
+
+// ============
+// === Kind ===
+// ============
+
+/// A module with marker types for each highlight kind. Those structs may implement various traits
+/// with operations specific for given highlight, for example [shape::AttrSetter] or
+/// [connected_entry::EndpointsGetter].
+#[allow(missing_docs, missing_copy_implementations)]
+pub mod kind {
+    #[derive(Debug)]
+    pub struct Selection;
+    #[derive(Debug)]
+    pub struct Hover;
+}
 
 
 
@@ -41,106 +60,58 @@ ensogl_core::define_endpoints_2! { <EntryParams: (frp::node::Data)>
         position(Vector2),
         contour(entry::Contour),
         color(color::Rgba),
+        /// Used in [Grid Views with headers](header::GridView). The y position of line separating
+        /// the header of scrolled down section from the rest of entries. If other entry than
+        /// this header is selected, the highlight should be top-clipped at this position.
+        header_separator(f32),
+        /// The y position of line clipping the highlight from the top. Usually it's just the top
+        /// of the viewport, but may be a lower line in case where a header in
+        /// [Grid Views with headers](header::GridView) should cover part of the highlight.
+        top_clip(f32),
         is_masked_layer_set(bool),
     }
 }
 
-/// A subset of [`entry::FRP`] endpoints used by the specific highlight [`Handler`].
-#[derive(Clone, CloneRef, Debug)]
-pub struct EntryEndpoints {
-    /// The "is_highlighted" input: may be `is_selected` or `is_hovered`.
-    flag:     frp::Any<bool>,
-    location: frp::Stream<(Row, Col)>,
-    contour:  frp::Stream<entry::Contour>,
-    color:    frp::Stream<color::Rgba>,
-}
 
 /// All highlight animations.
 #[derive(Clone, CloneRef, Debug)]
 pub struct Animations {
-    position:       Animation<Vector2>,
+    position_jump:  Animation<Vector2>,
+    position:       frp::Stream<Vector2>,
     size:           Animation<Vector2>,
     corners_radius: Animation<f32>,
     color:          Animation<color::Rgba>,
+    top_clip_jump:  Animation<f32>,
+    top_clip:       frp::Stream<f32>,
 }
 
 impl Animations {
     /// Set up the animations and connect their targets with handler's FRP.
     fn new<EntryParams: frp::node::Data>(frp: &Frp<EntryParams>) -> Self {
         let network = frp.network();
-        let position = Animation::<Vector2>::new(network);
+        let position_jump = Animation::<Vector2>::new(network);
         let size = Animation::<Vector2>::new(network);
         let corners_radius = Animation::<f32>::new(network);
         let color = Animation::<color::Rgba>::new(network);
+        let top_clip_jump = Animation::<f32>::new(network);
 
         frp::extend! { network
             init <- source_();
-            position.target <+ frp.position;
+            position <- all_with(&position_jump.value, &frp.position, |jump, pos| pos + jump);
             size.target <+ frp.contour.map(|&c| c.size);
             corners_radius.target <+ frp.contour.map(|&c| c.corners_radius);
             color.target <+ frp.color;
+            top_clip <- all_with(&top_clip_jump.value, &frp.top_clip, |jump, clip| clip + jump);
 
             // Skip animations when the highlight is not visible.
             size_val <- all(init, size.value)._1();
             not_visible <- size_val.map(|sz| sz.x < f32::EPSILON || sz.y < f32::EPSILON);
             corners_radius.skip <+ frp.color.gate(&not_visible).constant(());
-            position.skip <+ frp.position.gate(&not_visible).constant(());
+            position_jump.skip <+ frp.position.gate(&not_visible).constant(());
             color.skip <+ frp.color.gate(&not_visible).constant(());
         }
         init.emit(());
-        Self { position, size, corners_radius, color }
-    }
-}
-
-
-// ======================
-// === HighlightGuard ===
-// ======================
-
-/// A guard managing the connections between highlighted entry and the [`Handler`] FRP outputs.
-///
-/// Until dropped, this structure keeps connected the entry endpoints declaring the highlight
-/// appearance (`position`, `contour` and `color`) to the appropriate [`Handler`] endpoints.
-/// Also, the entry's flag (`is_selected` or `is_hovered`) will be set to `true` on construction and
-/// set back to `false` on drop.
-#[derive(Debug)]
-struct ConnectedEntryGuard {
-    network:           frp::Network,
-    /// An event emitted when we should drop this guard, for example when the Entry instance is
-    /// re-used in another location.
-    should_be_dropped: frp::Stream,
-    dropped:           frp::Source,
-}
-
-impl ConnectedEntryGuard {
-    /// Create guard for entry FRP at given location.
-    fn new_for_entry<EntryParams: frp::node::Data>(
-        entry_frp: EntryEndpoints,
-        row: Row,
-        col: Col,
-        frp: &api::private::Output<EntryParams>,
-    ) -> Self {
-        let network = frp::Network::new("HighlightedEntryGuard");
-        frp::extend! { network
-            init <- source_();
-            dropped <- source_();
-            contour <- all(init, entry_frp.contour)._1();
-            color <- all(init, entry_frp.color)._1();
-            entry_frp.flag <+ init.constant(true);
-            entry_frp.flag <+ dropped.constant(false);
-            frp.contour <+ contour;
-            frp.color <+ color;
-            location_change <- entry_frp.location.filter(move |loc| *loc != (row, col));
-            should_be_dropped <- location_change.constant(());
-        }
-        init.emit(());
-        Self { network, dropped, should_be_dropped }
-    }
-}
-
-impl Drop for ConnectedEntryGuard {
-    fn drop(&mut self) {
-        self.dropped.emit(());
+        Self { position_jump, position, size, corners_radius, color, top_clip_jump, top_clip }
     }
 }
 
@@ -150,100 +121,55 @@ impl Drop for ConnectedEntryGuard {
 // === Model ===
 // =============
 
-/// Function returning [`EntryEndpoints`] of given entry, used as parameter of highlight
-/// [`Handler`].
-pub type EntryEndpointsGetter<Entry> = fn(&Entry) -> EntryEndpoints;
-
 /// The inner data structure for [`Handler`].
 #[derive(Derivative)]
 #[derivative(Debug)]
-pub struct Data<Entry: 'static, EntryModel: frp::node::Data, EntryParams: frp::node::Data> {
-    app:                    Application,
-    grid:                   crate::GridViewTemplate<Entry, EntryModel, EntryParams>,
-    connected_entry:        Cell<Option<(Row, Col)>>,
-    guard:                  RefCell<Option<ConnectedEntryGuard>>,
-    output:                 api::private::Output<EntryParams>,
-    animations:             Animations,
+pub struct Data<InnerGridView, Entry> {
+    app:        Application,
+    grid:       InnerGridView,
+    animations: Animations,
     #[derivative(Debug = "ignore")]
-    entry_endpoints_getter: EntryEndpointsGetter<Entry>,
-    layers:                 RefCell<Option<layer::Handler<Entry, EntryModel, EntryParams>>>,
+    layers:     RefCell<Option<layer::Handler<InnerGridView>>>,
+    entry_type: PhantomData<Entry>,
 }
 
-impl<Entry, EntryModel, EntryParams> Data<Entry, EntryModel, EntryParams>
-where
-    EntryModel: frp::node::Data,
-    EntryParams: frp::node::Data,
+impl<InnerGridView, Entry> Data<InnerGridView, Entry>
+where InnerGridView: CloneRef
 {
-    fn new(
-        app: &Application,
-        grid: &crate::GridViewTemplate<Entry, EntryModel, EntryParams>,
-        output: &api::private::Output<EntryParams>,
-        animations: Animations,
-        entry_endpoints_getter: EntryEndpointsGetter<Entry>,
-    ) -> Self {
+    fn new(app: &Application, grid: &InnerGridView, animations: Animations) -> Self {
         Self {
             app: app.clone_ref(),
             grid: grid.clone_ref(),
-            connected_entry: default(),
-            guard: default(),
-            output: output.clone_ref(),
             animations,
-            entry_endpoints_getter,
             layers: default(),
+            entry_type: default(),
         }
-    }
-
-    fn drop_guard(&self) {
-        self.connected_entry.set(None);
-        self.guard.take();
     }
 }
 
-impl<E: Entry> Data<E, E::Model, E::Params> {
-    /// Drop old [`ConnectedEntryGuard`] and create a new one for new highlighted entry.
-    fn connect_new_highlighted_entry(self: &Rc<Self>, location: Option<(Row, Col)>) {
-        if let Some((row, col)) = location {
-            let current_entry = self.connected_entry.get();
-            let entry_changed = current_entry.map_or(true, |loc| loc != (row, col));
-            if entry_changed {
-                self.guard.take();
-                let entry = self.grid.get_entry(row, col);
-                *self.guard.borrow_mut() = entry.map(|e| {
-                    let endpoints = (self.entry_endpoints_getter)(&e);
-                    ConnectedEntryGuard::new_for_entry(endpoints, row, col, &self.output)
-                });
-                self.connected_entry.set(Some((row, col)));
-                self.set_up_guard_dropping();
-            }
-        } else {
-            self.drop_guard()
-        }
-    }
-
-    fn set_up_guard_dropping(self: &Rc<Self>) {
-        if let Some(guard) = &*self.guard.borrow() {
-            let network = &guard.network;
-            let this = self;
-            frp::extend! { network
-                eval_ guard.should_be_dropped (this.drop_guard());
-            }
-        }
-    }
-
-    fn setup_masked_layer(&self, layer: Option<&WeakLayer>) -> bool {
+impl<InnerGrid, E: Entry> Data<InnerGrid, E>
+where InnerGrid: AsRef<crate::GridView<E>>
+{
+    fn setup_masked_layer(
+        &self,
+        layer: Option<&WeakLayer>,
+        entries_params: frp::Stream<E::Params>,
+    ) -> bool
+    where
+        layer::Handler<InnerGrid>: layer::HasConstructor<InnerGrid = InnerGrid>,
+    {
         self.layers.take();
         let new_layers = layer.and_then(|l| l.upgrade()).map(|layer| {
-            let layers = layer::Handler::new(&self.app, &layer, &self.grid);
+            let layers = layer::Handler::<InnerGrid>::new(&self.app, &layer, &self.grid);
             let shape = &layers.shape;
-            let network = layers.grid.network();
-            let highlight_grid_frp = layers.grid.frp();
-            let frp = &self.output;
-            self.connect_with_shape::<HoverAttrSetter>(network, shape, false, None);
+            let network = layers.grid.as_ref().network();
+            let highlight_grid_frp = layers.grid.as_ref().frp();
+            self.connect_with_shape::<kind::Hover>(network, shape, false, None, None);
             frp::extend! { network
-                highlight_grid_frp.set_entries_params <+ frp.entries_params;
+                highlight_grid_frp.set_entries_params <+ entries_params;
             }
-            HoverAttrSetter::set_color(shape, color::Rgba::black());
-            SelectionAttrSetter::set_size(shape, default());
+            kind::Hover::set_color(shape, color::Rgba::black());
+            kind::Selection::set_size(shape, default());
             layers
         });
         let is_layer_set = new_layers.is_some();
@@ -262,13 +188,15 @@ impl<E: Entry> Data<E, E::Model, E::Params> {
         shape: &shape::View,
         connect_color: bool,
         hide: Option<&frp::Stream<bool>>,
+        bottom_clip: Option<&frp::Stream<f32>>,
     ) {
-        let grid_frp = self.grid.frp();
+        let grid_frp = self.grid.as_ref().frp();
         frp::extend! { network
             init <- source_();
-            position <- all(init, self.animations.position.value)._1();
+            position <- all(init, self.animations.position)._1();
             size <- all(init, self.animations.size.value)._1();
             corners_radius <- all(init, self.animations.corners_radius.value)._1();
+            top_clip <- all(init, self.animations.top_clip)._1();
         }
         let size = if let Some(hide) = hide {
             frp::extend! { network
@@ -283,11 +211,20 @@ impl<E: Entry> Data<E, E::Model, E::Params> {
             eval pos_and_viewport ([shape](&(pos, vp)) Setter::set_position(&shape, pos, vp));
             eval size ([shape](&size) Setter::set_size(&shape, size));
             eval corners_radius ([shape](&r) Setter::set_corners_radius(&shape, r));
+            top_clip_and_viewport <- all(top_clip, grid_frp.viewport);
+            eval top_clip_and_viewport ([shape](&(c, v)) Setter::set_top_clip(&shape, c, v));
         }
         if connect_color {
             frp::extend! { network
                 color <- all(init, self.animations.color.value)._1();
                 eval color ([shape](&color) Setter::set_color(&shape, color));
+            }
+        }
+        if let Some(bottom_clip) = bottom_clip {
+            frp::extend! {network
+                bottom_clip <- all(&init, bottom_clip)._1();
+                bottom_clip_and_viewport <- all(bottom_clip, grid_frp.viewport);
+                eval bottom_clip_and_viewport ([shape](&(c, v)) Setter::set_bottom_clip(&shape, c, v));
             }
         }
         init.emit(())
@@ -304,21 +241,102 @@ impl<E: Entry> Data<E, E::Model, E::Params> {
 ///
 /// This is a helper structure for [`selectable::GridView`] which handles a single highlight:
 /// selection or hover:
-/// * It provides an FRP API for given highlight, where the exact position, contour and color may be
-///   read, and the _Masked layer_ highlight mode may be configured (see the [grid
-///   view](`selectable::GridView`) documentation for more info about modes.
-/// * It can be connected to highlight shape, so it's position, contour and color will be updated.
+/// * It provides an FRP API for given highlight, where the exact position, contour, color and
+///   clipping may be read, and the _Masked layer_ highlight mode may be configured (see the [grid
+///   view](`selectable::GridView`) documentation for more info about modes).
+/// * It can be connected to highlight shape, so it's position, contour, color and clipping will be
+///   updated.
 /// * It sets/unsets the proper flag on [`Entry`] instance (`set_selected` or `set_hovered`).
+///
+/// The `Kind` parameter should be any from [`kind`] module.
 #[allow(missing_docs)]
 #[derive(CloneRef, Debug, Derivative, Deref)]
 #[derivative(Clone(bound = ""))]
-pub struct Handler<Entry: 'static, EntryModel: frp::node::Data, EntryParams: frp::node::Data> {
+pub struct Handler<Kind, InnerGridView, Entry, EntryParams: frp::node::Data> {
     #[deref]
-    pub frp: Frp<EntryParams>,
-    model:   Rc<Data<Entry, EntryModel, EntryParams>>,
+    pub frp:   Frp<EntryParams>,
+    model:     Rc<Data<InnerGridView, Entry>>,
+    kind_type: PhantomData<Kind>,
 }
 
-impl<E: Entry> Handler<E, E::Model, E::Params> {
+/// The handler of selection highlight.
+pub type SelectionHandler<InnerGridView, Entry, EntryParams> =
+    Handler<kind::Selection, InnerGridView, Entry, EntryParams>;
+
+/// The handler of mouse hover highlight.
+pub type HoverHandler<InnerGridView, Entry, EntryParams> =
+    Handler<kind::Hover, InnerGridView, Entry, EntryParams>;
+
+impl<InnerGridView, E: Entry> SelectionHandler<InnerGridView, E, E::Params>
+where InnerGridView: AsRef<crate::GridView<E>>
+{
+    /// Create selection highlight handler.
+    ///
+    /// The returned handler will have `entry_highlighted` properly connected to the passed `grid`.
+    pub fn new_connected(app: &Application, grid: &InnerGridView) -> Self
+    where Self: HasConstructor<InnerGridView = InnerGridView> {
+        let this = Self::new(app, grid);
+        this.entry_highlighted.attach(&grid.as_ref().entry_selected);
+        this
+    }
+}
+
+impl<InnerGridView, E: Entry> HoverHandler<InnerGridView, E, E::Params>
+where InnerGridView: AsRef<crate::GridView<E>>
+{
+    /// Create hover highlight handler.
+    ///
+    /// The returned handler will have `entry_highlighted` properly connected to the passed `grid`.
+    pub fn new_connected(app: &Application, grid: &InnerGridView) -> Self
+    where Self: HasConstructor<InnerGridView = InnerGridView> {
+        let this = Self::new(app, grid);
+        this.entry_highlighted.attach(&grid.as_ref().entry_hovered);
+        this
+    }
+}
+
+impl<Kind: shape::AttrSetter, InnerGridView, E: Entry> Handler<Kind, InnerGridView, E, E::Params>
+where InnerGridView: AsRef<crate::GridView<E>>
+{
+    /// Connects shape with the handler.
+    ///
+    /// The shape will be updated (using the specified `shape::AttrSetter`) according to the
+    /// `position`, `contour`, `color` and `top_clip` outputs.
+    pub fn connect_with_shape(&self, shape: &shape::View) {
+        let network = self.frp.network();
+        let shape_hidden = (&self.frp.is_masked_layer_set).into();
+        self.model.connect_with_shape::<Kind>(network, shape, true, Some(&shape_hidden), None);
+    }
+}
+
+impl<Kind: shape::AttrSetter, E: Entry, HeaderEntry: Entry>
+    Handler<Kind, header::GridView<E, HeaderEntry>, E, E::Params>
+{
+    /// Connects shape designed for highlighting headers with the handler.
+    ///
+    /// The shape will be updated (using the specified `shape::AttrSetter`) according to the
+    /// `position`, `contour`, `color` and `top_clip` outputs, and it's bottom clip will be always
+    /// set to the current column's pushed-down header bottom (the `header_separator` value of
+    /// [`header::FRP`]).
+    pub fn connect_with_header_shape(&self, shape: &shape::View) {
+        let network = self.frp.network();
+        let shape_hidden = (&self.frp.is_masked_layer_set).into();
+        let bottom_clip = (&self.frp.header_separator).into();
+        self.model.connect_with_shape::<Kind>(
+            network,
+            shape,
+            true,
+            Some(&shape_hidden),
+            Some(&bottom_clip),
+        );
+    }
+}
+
+/// A trait implemented by every [`Handler`] able to be constructed.
+pub trait HasConstructor {
+    /// The exact type of the _inner_ Grid View.
+    type InnerGridView;
+
     /// Create new Handler.
     ///
     /// This is a generic constructor: the exact entry endpoints handled by it should be specified
@@ -327,88 +345,188 @@ impl<E: Entry> Handler<E, E::Model, E::Params> {
     ///
     /// Use [`new_for_selection_connected`] or [`new_for_hover_connected`] to create a fully working
     /// handler for specific highlight.
-    pub fn new(
-        app: &Application,
-        grid: &crate::GridView<E>,
-        entry_endpoints_getter: EntryEndpointsGetter<E>,
-    ) -> Self {
+    fn new(app: &Application, grid: &Self::InnerGridView) -> Self;
+}
+
+impl<Kind: EndpointsGetter, E: Entry> HasConstructor
+    for Handler<Kind, crate::GridView<E>, E, E::Params>
+{
+    type InnerGridView = crate::GridView<E>;
+
+    fn new(app: &Application, grid: &Self::InnerGridView) -> Self {
         let frp = Frp::new();
-        let animations = Animations::new(&frp);
-        let out = &frp.private.output;
-        let model = Rc::new(Data::new(app, grid, out, animations, entry_endpoints_getter));
         let network = frp.network();
+        let animations = Animations::new(&frp);
+        let entry_getter =
+            f!((r, c) grid.get_entry(r, c).map(|e| Kind::get_endpoints_from_entry(&e)));
+        let connected_entry_out = connected_entry::Output::new(network);
+        let connected_entry = ConnectedEntry::new(entry_getter, connected_entry_out.clone_ref());
+        let model = Rc::new(Data::new(app, grid, animations.clone_ref()));
+        let out = &frp.private.output;
         let grid_frp = grid.frp();
         frp::extend! {network
+
+            // === Updating `connected_entry` Field ===
             shown_with_highlighted <-
                 grid_frp.entry_shown.map2(&frp.entry_highlighted, |a, b| (*a, *b));
             highlighted_is_shown <-
-                shown_with_highlighted.map(|(sh, hlt)| hlt.contains(sh).and_option(*hlt));
-            should_reconnect <- any(frp.entry_highlighted, highlighted_is_shown);
-            eval should_reconnect ((loc) model.connect_new_highlighted_entry(*loc));
+                shown_with_highlighted.filter_map(|(sh, hlt)| hlt.contains(sh).and_option(Some(*hlt)));
+            should_reconnect_entry <- any(frp.entry_highlighted, highlighted_is_shown);
+            eval should_reconnect_entry ((loc) connected_entry.connect_new_highlighted_entry(*loc));
+
+
+            // === Highlight Position ===
 
             became_highlighted <- frp.entry_highlighted.filter_map(|l| *l);
-            let column_widths = model.grid.model().column_widths.clone_ref();
-            out.position <+ became_highlighted.all_with(
-                &grid_frp.entries_size,
-                move |&(row, col), &es| entry_position(row, col, es, &column_widths)
+            position_after_highlight <- became_highlighted.map(
+                f!((&(row, col)) model.grid.entry_position(row, col))
             );
-            none_highlightd <- frp.entry_highlighted.filter(|opt| opt.is_none()).constant(());
-            out.contour <+ none_highlightd.constant(default());
+            out.position <+ position_after_highlight;
+            prev_position <- out.position.previous();
+            new_jump <- position_after_highlight.map2(&prev_position, |pos, prev| prev - pos);
+            animations.position_jump.target <+ new_jump;
+            animations.position_jump.skip <+ new_jump.constant(());
+            animations.position_jump.target <+ new_jump.constant(Vector2::zero());
+
+
+            // === Color and Contour ===
+
+            out.contour <+ connected_entry_out.contour;
+            out.color <+ connected_entry_out.color;
+
+            none_highlighted <- frp.entry_highlighted.filter(|opt| opt.is_none()).constant(());
+            out.contour <+ none_highlighted.constant(default());
+
+
+            // === Setting up Masked Layer ===
 
             out.entries_params <+ frp.set_entries_params;
+            let entries_params = out.entries_params.clone_ref();
             out.is_masked_layer_set <+
-                frp.setup_masked_layer.map(f!((layer) model.setup_masked_layer(layer.as_ref())));
-
+                frp.setup_masked_layer.map(f!([model, entries_params](layer) model.setup_masked_layer(layer.as_ref(), (&entries_params).into())));
         }
 
-        Self { frp, model }
+        Self { frp, model, kind_type: default() }
     }
+}
 
-    /// Create selection highlight handler.
-    ///
-    /// The returned handler will have `entry_highlighted` properly connected to the passed `grid`.
-    pub fn new_for_selection_connected(app: &Application, grid: &crate::GridView<E>) -> Self {
-        let this = Self::new(app, grid, Self::selection_endpoints_getter);
-        this.entry_highlighted.attach(&grid.entry_selected);
-        this
-    }
+impl<Kind: EndpointsGetter, E: Entry, HeaderEntry: Entry<Params = E::Params>> HasConstructor
+    for Handler<Kind, header::GridView<E, HeaderEntry>, E, E::Params>
+{
+    type InnerGridView = header::GridView<E, HeaderEntry>;
 
-    /// Create hover highlight handler.
-    ///
-    /// The returned handler will have `entry_highlighted` properly connected to the passed `grid`.
-    pub fn new_for_hover_connected(app: &Application, grid: &crate::GridView<E>) -> Self {
-        let this = Self::new(app, grid, Self::hover_endpoints_getter);
-        this.entry_highlighted.attach(&grid.entry_hovered);
-        this
-    }
+    fn new(app: &Application, grid: &Self::InnerGridView) -> Self {
+        let frp = Frp::new();
+        let network = frp.network();
+        let animations = Animations::new(&frp);
+        let entry_getter =
+            f!((r, c) grid.get_entry(r, c).map(|e| Kind::get_endpoints_from_entry(&e)));
+        let header_getter =
+            f!((r, c) grid.get_header(r, c).map(|e| Kind::get_endpoints_from_entry(&e)));
+        let connected_entry_out = connected_entry::Output::new(network);
+        let connected_header_out = connected_entry::Output::new(network);
+        let connected_entry = ConnectedEntry::new(entry_getter, connected_entry_out.clone_ref());
+        let connected_header = ConnectedEntry::new(header_getter, connected_header_out.clone_ref());
+        let model = Rc::new(Data::new(app, grid, animations.clone_ref()));
+        let out = &frp.private.output;
+        let grid_frp = grid.frp();
+        let headers_frp = grid.header_frp();
+        frp::extend! {network
 
-    fn selection_endpoints_getter(entry: &E) -> EntryEndpoints {
-        let frp = entry.frp();
-        EntryEndpoints {
-            flag:     frp.set_selected.clone_ref(),
-            location: frp.set_location.clone_ref().into(),
-            contour:  frp.contour.clone_ref().into(),
-            color:    frp.selection_highlight_color.clone_ref().into(),
+
+            // === Updating `connected_entry` Field ===
+
+            shown_with_highlighted <-
+                grid_frp.entry_shown.map2(&frp.entry_highlighted, |a, b| (*a, *b));
+            highlighted_is_shown <-
+                shown_with_highlighted.filter_map(|(sh, hlt)| hlt.contains(sh).and_option(Some(*hlt)));
+            should_reconnect_entry <- any(frp.entry_highlighted, highlighted_is_shown);
+            eval should_reconnect_entry ((loc) connected_entry.connect_new_highlighted_entry(*loc));
+
+
+            // === Updating `connected_header` Field ===
+
+            header_shown_with_highlighted <-
+                headers_frp.header_shown.map2(&frp.entry_highlighted, |a, b| (*a, *b));
+            highlighted_header_is_shown <-
+                header_shown_with_highlighted.filter_map(|(sh, hlt)| hlt.contains(sh).and_option(Some(*hlt)));
+            should_reconnect_header <- any(frp.entry_highlighted, highlighted_header_is_shown);
+            eval should_reconnect_header ((loc) connected_header.connect_new_highlighted_entry(*loc));
+            header_hidden_with_highlighted <-
+                headers_frp.header_hidden.map2(&frp.entry_highlighted, |a, b| (*a, *b));
+            should_disconnect_header <- header_hidden_with_highlighted.filter_map(|(hd, hlt)| hlt.contains(hd).and_option(*hlt));
+            eval_ should_disconnect_header (connected_header.drop_guard());
+
+
+            // === Highlight Position ===
+
+            became_highlighted <- frp.entry_highlighted.filter_map(|l| *l);
+            position_after_highlight <- became_highlighted.map(
+                f!((&(row, col)) model.grid.header_or_entry_position(row, col))
+            );
+            position_after_header_disconnect <- should_disconnect_header.map(
+                f!((&(row, col)) model.grid.header_or_entry_position(row, col))
+            );
+            highligthed_header_pos_change <- headers_frp.header_position_changed.map2(
+                &frp.entry_highlighted,
+                |&(row, col, pos), h| h.contains(&(row, col)).as_some(pos)
+            );
+            highligthed_header_pos_change <- highligthed_header_pos_change.filter_map(|p| *p);
+            out.position <+ position_after_highlight;
+            out.position <+ position_after_header_disconnect;
+            out.position <+ highligthed_header_pos_change;
+            prev_position <- out.position.previous();
+            new_jump <- position_after_highlight.map2(&prev_position, |pos, prev| prev - pos);
+            animations.position_jump.target <+ new_jump;
+            animations.position_jump.skip <+ new_jump.constant(());
+            animations.position_jump.target <+ new_jump.constant(Vector2(0.0, 0.0));
+
+
+            // === Contour and Color ===
+
+            let header_connected = &connected_header_out.is_entry_connected;
+            out.contour <+ all_with3(
+                header_connected,
+                &connected_entry_out.contour,
+                &connected_header_out.contour,
+                |&from_header, &entry, &header| if from_header {header} else {entry}
+            );
+            out.color <+ all_with3(
+                header_connected,
+                &connected_entry_out.color,
+                &connected_header_out.color,
+                |&from_header, &entry, &header| if from_header {header} else {entry}
+            );
+
+            none_highlighted <- frp.entry_highlighted.filter(|opt| opt.is_none()).constant(());
+            out.contour <+ none_highlighted.constant(default());
+
+
+            // === Setting up Masked Layer ===
+
+            out.entries_params <+ frp.set_entries_params;
+            let entries_params = out.entries_params.clone_ref();
+            out.is_masked_layer_set <+
+                frp.setup_masked_layer.map(f!([model, entries_params](layer) model.setup_masked_layer(layer.as_ref(), (&entries_params).into())));
+
+
+            // === Highlight Shape Clipping ===
+
+            highlight_column <- became_highlighted._1();
+            header_moved_in_highlight_column <- headers_frp.header_position_changed.map2(
+                &frp.entry_highlighted,
+                |&(_, col, _), h| h.map_or(false, |(_, h_col)| col == h_col)
+            ).filter(|v| *v).constant(());
+            out.header_separator <+ all_with(&highlight_column, &header_moved_in_highlight_column, f!((col, ()) model.grid.header_separator(*col)));
+            new_top_clip <- all_with3(&out.header_separator, header_connected, &grid_frp.viewport, |&sep, &hc, v| if hc {v.top} else {sep});
+            out.top_clip <+ new_top_clip;
+            prev_top_clip <- out.top_clip.previous();
+            new_clip_jump <- new_top_clip.sample(&became_highlighted).map2(&prev_top_clip, |c, p| p - c);
+            animations.top_clip_jump.target <+ new_clip_jump;
+            animations.top_clip_jump.skip <+ new_clip_jump.constant(());
+            animations.top_clip_jump.target <+ new_clip_jump.constant(0.0);
         }
-    }
 
-    fn hover_endpoints_getter(entry: &E) -> EntryEndpoints {
-        let frp = entry.frp();
-        EntryEndpoints {
-            flag:     frp.set_hovered.clone_ref(),
-            location: frp.set_location.clone_ref().into(),
-            contour:  frp.contour.clone_ref().into(),
-            color:    frp.hover_highlight_color.clone_ref().into(),
-        }
-    }
-
-    /// Connects shape with the handler.
-    ///
-    /// The shape will be updated (using the specified `shape::AttrSetter`) according to the
-    /// `position`, `contour` and `color` outputs.
-    pub fn connect_with_shape<Setter: shape::AttrSetter>(&self, shape: &shape::View) {
-        let network = self.frp.network();
-        let shape_hidden = (&self.frp.is_masked_layer_set).into();
-        self.model.connect_with_shape::<Setter>(network, shape, true, Some(&shape_hidden));
+        Self { frp, model, kind_type: default() }
     }
 }